[apps] Operations in double when precision required

+++ apps/probability/calculation/calculation.h
@@ -20,15 +20,15 @@ public:
   virtual int numberOfParameters() = 0;
   virtual int numberOfEditableParameters();
   virtual I18n::Message legendForParameterAtIndex(int index) = 0;
-  virtual void setParameterAtIndex(float f, int index) = 0;
-  virtual float parameterAtIndex(int index) = 0;
-  virtual float lowerBound();
-  virtual float upperBound();
+  virtual void setParameterAtIndex(double f, int index) = 0;
+  virtual double parameterAtIndex(int index) = 0;
+  virtual double lowerBound();
+  virtual double upperBound();
 protected:
   /* Parameters in probability application are rounded to 3 decimals. This is
    * due to the limited precision of some calculation (e. g. standard normal
    * cumulative distributive function or inverse). */
-  constexpr static float k_precision = 0.001f;
+  constexpr static double k_precision = 0.001;
   virtual void compute(int indexKnownElement) = 0;
   Law * m_law;
 };
