+++ apps/probability/calculation/left_integral_calculation.h
@@ -11,13 +11,13 @@ public:
   Type type() override;
   int numberOfParameters() override;
   I18n::Message legendForParameterAtIndex(int index) override;
-  void setParameterAtIndex(float f, int index) override;
-  float parameterAtIndex(int index) override;
-  float upperBound() override;
+  void setParameterAtIndex(double f, int index) override;
+  double parameterAtIndex(int index) override;
+  double upperBound() override;
 private:
   void compute(int indexKnownElement) override;
-  float m_upperBound;
-  float m_result;
+  double m_upperBound;
+  double m_result;
 };
 
 }
