+++ apps/shared/curve_view_cursor.h
@@ -6,14 +6,14 @@ namespace Shared {
 class CurveViewCursor {
 public:
   CurveViewCursor();
-  float x();
-  float y();
-  void moveTo(float x, float y);
+  double x();
+  double y();
+  void moveTo(double x, double y);
 private:
-  static float clipped(float f, bool canBeInfinite);
-  constexpr static float k_maxFloat = 1E+8f;
-  float m_x;
-  float m_y;
+  static double clipped(double f, bool canBeInfinite);
+  constexpr static double k_maxFloat = 1E+8;
+  double m_x;
+  double m_y;
 };
 
 }
