+++ apps/shared/go_to_parameter_controller.h
@@ -14,7 +14,7 @@ public:
   int numberOfRows() override;
   bool handleEvent(Ion::Events::Event event) override;
 protected:
-  constexpr static float k_maxDisplayableFloat = 1E8f;
+  constexpr static double k_maxDisplayableFloat = 1E8;
   CurveViewCursor * m_cursor;
   InteractiveCurveViewRange * m_graphRange;
 private:
